新增物品，方块

+++ src/main/java/com/arc/outland_horizon/ModArmorMaterials.java
@@ -18,21 +18,21 @@ public class ModArmorMaterials {
     public static ArmorMaterial CONDENSE = createArmorMaterial("condense", ModTiers.CONDENSE_LEVEL,"condense");
     public static ArmorMaterial VOID = createArmorMaterial("void", ModTiers.VOID_CRYSTAL_LEVEL,"void");
     */
-    public static @NotNull ArmorMaterial createArmorMaterial(String name, float level, Ingredient repairIngredient) {
+    public static @NotNull ArmorMaterial createArmorMaterial(String name, double level, Ingredient repairIngredient) {
         return new ArmorMaterial() {
             @Override
             public int getDurabilityForType(ArmorItem.@NotNull Type type) {
-                return new int[]{13, 18, 25, 11}[type.getSlot().getIndex()] * Math.round(14 * level);
+                return (new int[]{13, 18, 25, 11}[type.getSlot().getIndex()] * (int) Math.round(14 * level));
             }
 
             @Override
             public int getDefenseForType(ArmorItem.@NotNull Type type) {
-                return new int[]{Math.round(1.5f * level), Math.round(3 * level), Math.round(4 * level), Math.round(1.5f * level)}[type.getSlot().getIndex()];
+                return new int[]{(int) Math.round(1.5f * level), (int) Math.round(3 * level), (int) Math.round(4 * level), (int) Math.round(1.5f * level)}[type.getSlot().getIndex()];
             }
 
             @Override
             public int getEnchantmentValue() {
-                return Math.round(9 * level);
+                return (int) Math.round(9 * level);
             }
 
             @Override
@@ -52,12 +52,12 @@ public class ModArmorMaterials {
 
             @Override
             public float getToughness() {
-                return level >= 2 ? level : 0f;
+                return level >= 2 ? (float) level : 0f;
             }
 
             @Override
             public float getKnockbackResistance() {
-                return level >= 3 ? 0.1f * level / 3.0f : 0f;
+                return level >= 3 ? (0.1f * (float) level / 3.0f) : 0f;
             }
         };
     }
